task1/backend/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.34
